requirements.txt
openai
google-generativeai
tree-sitter
tree-sitter-python
tree-sitter-c
tree-sitter-cpp
chromadb
pydantic
python-dotenv
